Revert "Merge branch 'wishradio_hosted' of https://github.com/saumishr/apps into sub-menu-categories"
This reverts commit a8a2feedbb97e3d9d38b01e2901aebd755fcf182, reversing
changes made to ec086387838fe49b2bdad9374365677e2576b468.

+++ userProfile/views.py
@@ -40,9 +40,7 @@ import datetime
 import os
 import uuid
 import json
-
 from django.core.urlresolvers import reverse
-from storages.backends.s3boto import S3BotoStorage
 import StringIO
 from django.core.files.uploadedfile import InMemoryUploadedFile
 from django.views.decorators.csrf import csrf_exempt
@@ -1028,18 +1026,9 @@ def shareObject(request, content_type_id, object_id, ):
 	else:
 		raise Http404();
 
-def deleteFileS3(file):
-	"""
-	We cannot use file.path as its not yet implemented in S3BotoStorage backend and threfore throws an exception.
-	file.url returns complete web url. We need to get the path after media/ as bucket/static/media is storage root.
-	"""
-	absolute_url = file.url
-	paths = absolute_url.split('media/')
-	if len(paths) > 0:
-		path = paths[1]
-		storage = S3BotoStorage(location=settings.STORAGE_ROOT)
-		if storage.exists(path):
-			storage.delete(path)
+def deleteFile(file):
+    storage, path = file.storage, file.path
+    storage.delete(path)
 
 def deleteObject(request, content_type_id, object_id ):
     error_codes = []
@@ -1137,12 +1126,12 @@ def deleteObject(request, content_type_id, object_id ):
                 OptionalReviewRatingObj.delete()
 
         elif isinstance(object, GenericWish):
-            try:
-                deleteFileS3(object.wishimage.image)
-            except:
-                pass
-            
-            object.wishimage.delete()
+            if object.wishimage:
+                try:
+                    deleteFile(object.wishimage.image)
+                except:
+                    pass
+                object.wishimage.delete()
 
         """
             Finally nuke the actual object.
@@ -1182,21 +1171,6 @@ def get_object_owner_helper(content_type_id, object_id):
 			
 	return owner
 
-def save_file_s3(file, path=''):
-    ''' Little helper to save a file
-    '''
-    filename = file._get_name()
-    new_filename =  u'{name}.{ext}'.format(  name=uuid.uuid4().hex,
-                                             ext=os.path.splitext(filename)[1].strip('.'))
-
-    dir_path =  str(path) #'%s/%s' % (settings.MEDIA_URL, str(path))
-
-    save_path = os.path.join(dir_path, new_filename)
-    storage=S3BotoStorage(location=settings.STORAGE_ROOT)
-    storage.save(save_path, file)
-
-    return save_path
-
 def save_file(file, path=''):
     ''' Little helper to save a file
     '''
@@ -1204,7 +1178,7 @@ def save_file(file, path=''):
     new_filename =  u'{name}.{ext}'.format(  name=uuid.uuid4().hex,
                                              ext=os.path.splitext(filename)[1].strip('.'))
 
-    dir_path =  '%s/%s' % (settings.MEDIA_URL, str(path))
+    dir_path =  '%s/%s' % (settings.MEDIA_ROOT, str(path))
 
     if not os.path.exists(dir_path):
     	os.makedirs(dir_path)
@@ -1241,19 +1215,17 @@ def edit_blog_image(request, blogpost_id):
 				featuredImageObj = request.FILES['featured_image']
 				if featuredImageObj:
 					new_file_rel_path = 'users/store/%s/images/' % (blogpost.id)
-					new_file_path = save_file_s3(featuredImageObj, new_file_rel_path)
+					new_file_path = save_file(featuredImageObj, new_file_rel_path)
 					if blogpost.featured_image:
-						old_file_path = blogpost.featured_image.path #'%s/%s' % (settings.MEDIA_URL, str(blogpost.featured_image.path))
-						storage = S3BotoStorage(location=settings.STORAGE_ROOT)
-						if storage.exists(old_file_path):
-							storage.delete(old_file_path)
-						
+						old_file_path = '%s/%s' % (settings.MEDIA_ROOT, str(blogpost.featured_image.path))
+						if os.path.exists(old_file_path):
+							os.remove(old_file_path)
+
 					blogpost.featured_image = new_file_path
 					blogpost.save()
 
 				if request.is_ajax():
-					url = os.path.join(settings.MEDIA_URL, blogpost.featured_image.path)
-					return HttpResponse(simplejson.dumps(dict(success=True, image_url=url)))
+					return HttpResponse(simplejson.dumps(dict(success=True, image_url=blogpost.featured_image.url)))
 				else:
 					return HttpResponseRedirect(blogpost.get_absolute_url())
